[#333] extension 파일 분리
