+++ Mohaeng/Mohaeng/Sources/ViewControllers/Feed/WritingViewController.swift
@@ -140,7 +140,7 @@ class WritingViewController: UIViewController {
     // MARK: - Functions
     private func makeRoundedViews() {
         yellowBackgroundView.makeRounded(radius: 20)
-        doneButton.makeRoundedSpecificCorner(corners: [.topLeft,.topRight], cornerRadius: 30.0)
+        doneButton.makeRoundedSpecificCorner(corners: [.topLeft, .topRight], cornerRadius: 30.0)
     }
     
     private func initViewContoller() {
@@ -236,22 +236,7 @@ class WritingViewController: UIViewController {
     
 }
 
-extension UIImage {
-    func resized(to size: CGSize) -> UIImage {
-        return UIGraphicsImageRenderer(size: size).image { _ in
-            draw(in: CGRect(origin: .zero, size: size))
-        }
-    }
-}
 
-extension UIButton {
-    func alignTextLeft(spacing: CGFloat = 8.0) {
-        contentHorizontalAlignment = .leading
-        titleEdgeInsets = UIEdgeInsets(top: 0, left: spacing/2, bottom: 0, right: -spacing/2)
-        imageEdgeInsets = UIEdgeInsets(top: 0, left: -spacing/2, bottom: 0, right: spacing/2)
-        contentEdgeInsets = UIEdgeInsets(top: 0, left: spacing / 2, bottom: 0, right: -spacing / 2)
-    }
-}
 
 extension WritingViewController {
     @objc
@@ -430,16 +415,3 @@ extension WritingViewController {
     }
 }
 
-extension UIButton {
-    func setBackgroundColor(_ color: UIColor, for state: UIControl.State) {
-        UIGraphicsBeginImageContext(CGSize(width: 1.0, height: 1.0))
-        guard let context = UIGraphicsGetCurrentContext() else { return }
-        context.setFillColor(color.cgColor)
-        context.fill(CGRect(x: 0.0, y: 0.0, width: 1.0, height: 1.0))
-        
-        let backgroundImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-         
-        self.setBackgroundImage(backgroundImage, for: state)
-    }
-}
